Front end validation on form
This now has some front end validation on the form for record request.
As we spoke about on Thursday this is going to get moved to the page
where we have the list of all our records (My Records?) where we can
then submit the form via the modal. So this stuff I'm adding is going to
get moved around, but the validation is there now.